Move eos repo
