+++ lib/datasource/remote/api/invoice_repository.dart
@@ -3,7 +3,7 @@ import 'package:async/async.dart';
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:dart_esr/dart_esr.dart' as esr;
 import 'package:seeds/datasource/local/models/token_data_model.dart';
-import 'package:seeds/datasource/remote/api/eos_repository.dart';
+import 'package:seeds/datasource/remote/api/eos_repo/eos_repository.dart';
 import 'package:seeds/datasource/remote/firebase/firebase_remote_config.dart';
 import 'package:seeds/domain-shared/app_constants.dart';
