+++ lib/datasource/remote/api/send_eos_transaction_repository.dart
@@ -2,7 +2,7 @@ import 'package:async/async.dart';
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:eosdart/eosdart.dart';
 import 'package:seeds/datasource/local/models/eos_transaction.dart';
-import 'package:seeds/datasource/remote/api/eos_repository.dart';
+import 'package:seeds/datasource/remote/api/eos_repo/eos_repository.dart';
 
 class SendTransactionRepository extends EosRepository {
   Future<Result> sendTransaction({
