+++ lib/screens/explore_screens/vote_screens/delegate/delegates_tab/interactor/usecase/delegates_load_data_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:async/async.dart';
 import 'package:seeds/datasource/local/settings_storage.dart';
-import 'package:seeds/datasource/remote/api/eos_repository.dart';
+import 'package:seeds/datasource/remote/api/eos_repo/eos_repository.dart';
 import 'package:seeds/datasource/remote/api/members_repository.dart';
 import 'package:seeds/datasource/remote/api/proposals_repository.dart';
 import 'package:seeds/datasource/remote/model/delegate_model.dart';
