+++ lib/screens/explore_screens/vote_screens/delegate/delegates_tab/interactor/usecase/has_delegate_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:seeds/datasource/local/settings_storage.dart';
-import 'package:seeds/datasource/remote/api/eos_repository.dart';
+import 'package:seeds/datasource/remote/api/eos_repo/eos_repository.dart';
 import 'package:seeds/datasource/remote/api/proposals_repository.dart';
 import 'package:seeds/datasource/remote/model/delegate_model.dart';
