+++ lib/screens/explore_screens/vote_screens/vote/interactor/usecases/get_all_delegates_use_case.dart
@@ -1,5 +1,5 @@
 import 'package:seeds/datasource/local/settings_storage.dart';
-import 'package:seeds/datasource/remote/api/eos_repository.dart';
+import 'package:seeds/datasource/remote/api/eos_repo/eos_repository.dart';
 import 'package:seeds/datasource/remote/api/proposals_repository.dart';
 import 'package:seeds/domain-shared/result_to_state_mapper.dart';
